requirements.txt
accelerate==0.21.0
controlnet-aux==0.0.7
diffusers==0.24.0
einops==0.4.1
gradio==4.26.0
imageio==2.33.0
imageio[ffmpeg]
matplotlib==3.8.3
numpy==1.23.5
omegaconf==2.2.3
Pillow==9.5.0
safetensors==0.4.2
onnxruntime-gpu==1.13.1
open-clip-torch==2.20.0
opencv-contrib-python-headless==4.8.1.78
opencv-python-headless==4.8.1.78
pyrender==0.1.45
torchmetrics==1.2.1
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.1
transformers==4.30.2
xformers==0.0.25.post1
ultralytics==8.1.34
